requirements.txt
attrs==17.4.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JWT-Simple==0.0.3
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
-e git+git@github.com:a-rich/LiquiDemocracy.git@4f0fa0b2ac996c092fc5f047684373507edb5914#egg=liquidemocracy&subdirectory=project/server
lxml==4.2.0
MarkupSafe==1.0
mongoengine==0.15.0
nltk==3.2.5
numpy==1.14.2
pandas==0.22.0
pdfminer3k==1.3.1
pluggy==0.6.0
ply==3.11
py==1.5.2
PyJWT==1.5.3
pymongo==3.6.0
pytest==3.4.1
pytest-flask==0.10.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
slate==0.5.2
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
